backend/pyproject.toml
[project]
name = "citizen-service-backend"
version = "0.1.0"
description = "FastAPI backend for citizen service marketplace"
authors = [
  { name = "Final Year Project Team" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.111.0,<0.112.0",
  "uvicorn[standard]>=0.30.0,<0.31.0",
  "sqlmodel>=0.0.22,<0.0.23",
  "alembic>=1.13.1,<1.14.0",
  "passlib[bcrypt]>=1.7.4,<1.8.0",
  "pyjwt[crypto]>=2.8.0,<2.9.0",
  "python-multipart>=0.0.9,<0.1.0",
  "pydantic-settings>=2.2.1,<2.3.0",
  "email-validator>=2.1.1,<2.2.0",
  "redis>=5.0.4,<5.1.0",
  "celery>=5.4.0,<5.5.0",
  "httpx>=0.27.0,<0.28.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.1,<8.3.0",
  "pytest-asyncio>=0.23.6,<0.24.0",
  "ruff>=0.4.5,<0.5.0",
  "mypy>=1.8.0,<1.9.0"
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

